compact the tests in the proper folder
