+++ sdk/core/azure-core/test/perf/CMakeLists.txt
@@ -9,8 +9,12 @@ set(CMAKE_CXX_STANDARD_REQUIRED True)
 
 set(
   AZURE_CORE_PERF_TEST_HEADER
+  inc/azure/core/test/delay_test.hpp
+  inc/azure/core/test/exception_test.hpp
+  inc/azure/core/test/extended_options_test.hpp
   inc/azure/core/test/http_transport_test.hpp
   inc/azure/core/test/json_test.hpp
+  inc/azure/core/test/no_op_test.hpp
   inc/azure/core/test/nullable_test.hpp
   inc/azure/core/test/pipeline_test.hpp
   inc/azure/core/test/uuid_test.hpp
