+++ sdk/core/azure-core/test/perf/inc/azure/core/test/extended_options_test.hpp
@@ -3,7 +3,7 @@
 
 /**
  * @file
- * @brief Define an empty test.
+ * @brief An example of a performance test that defines a test option.
  *
  */
 
@@ -12,28 +12,30 @@
 #include <azure/perf.hpp>
 
 #include <memory>
+#include <vector>
 
-namespace Azure { namespace Perf { namespace Test {
+namespace Azure { namespace Core { namespace Test {
 
   /**
-   * @brief The no op test is an empty test used to measure the performance framework alone.
+   * @brief A performance test that defines a test option.
    *
    */
-  class NoOp : public Azure::Perf::PerfTest {
+  class ExtendedOptionsTest : public Azure::Perf::PerfTest {
   public:
     /**
-     * @brief Construct a new No Op test.
+     * @brief Construct a new Extended Options Test object.
      *
-     * @param options The test options.
+     * @param options The command-line parsed options.
      */
-    NoOp(Azure::Perf::TestOptions options) : PerfTest(options) {}
+    ExtendedOptionsTest(Azure::Perf::TestOptions options) : PerfTest(options) {}
 
     /**
-     * @brief Define an empty test.
+     * @brief The test definition
      *
      */
     void Run(Azure::Core::Context const&) override
-    { // Get the option or a default value of 0
+    {
+      // Get the option or a default value of 0
       auto myTestOption = m_options.GetOptionOrDefault("extraOption", 0);
       (void)myTestOption;
     }
@@ -56,12 +58,13 @@ namespace Azure { namespace Perf { namespace Test {
     static Azure::Perf::TestMetadata GetTestMetadata()
     {
       return {
-          "NoOp",
-          "Simplest test to measure the performance framework speed.",
+          "extendedOptions",
+          "Demostrate how to include a test option to a test and measures how expensive is to do "
+          "it.",
           [](Azure::Perf::TestOptions options) {
-            return std::make_unique<Azure::Perf::Test::NoOp>(options);
+            return std::make_unique<Azure::Core::Test::ExtendedOptionsTest>(options);
           }};
     }
   };
 
-}}} // namespace Azure::Perf::Test
+}}} // namespace Azure::Core::Test
