+++ sdk/core/azure-core/test/perf/src/azure_core_perf_test.cpp
@@ -1,8 +1,12 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+#include "azure/core/test/delay_test.hpp"
+#include "azure/core/test/exception_test.hpp"
+#include "azure/core/test/extended_options_test.hpp"
 #include "azure/core/test/http_transport_test.hpp"
 #include "azure/core/test/json_test.hpp"
+#include "azure/core/test/no_op_test.hpp"
 #include "azure/core/test/nullable_test.hpp"
 #include "azure/core/test/pipeline_test.hpp"
 #include "azure/core/test/uuid_test.hpp"
@@ -16,8 +20,12 @@ int main(int argc, char** argv)
 
   // Create the test list
   std::vector<Azure::Perf::TestMetadata> tests{
+      Azure::Core::Test::DelayTest::GetTestMetadata(),
+      Azure::Core::Test::ExceptionTest::GetTestMetadata(),
+      Azure::Core::Test::ExtendedOptionsTest::GetTestMetadata(),
       Azure::Core::Test::HTTPTransportTest::GetTestMetadata(),
       Azure::Core::Test::JsonTest::GetTestMetadata(),
+      Azure::Core::Test::NoOp::GetTestMetadata(),
       Azure::Core::Test::NullableTest::GetTestMetadata(),
       Azure::Core::Test::PipelineTest::GetTestMetadata(),
       Azure::Core::Test::UuidTest::GetTestMetadata()};
